requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
asgiref==3.10.0
asyncio==3.4.3
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
crispy-bootstrap5==2025.4
disutils==1.4.32.post2
Django==5.2.7
django-allauth==65.4.1
django-ckeditor==6.7.2
django-ckeditor-5==0.2.18
django-crispy-forms==2.3
django-debug-toolbar==5.0.1
django-environ==0.12.0
django-jalali-date==2.0.0
django-js-asset==3.1.2
django-mptt==0.17.0
django-rosetta==0.10.2
django-widget-tweaks==1.5.0
djangorestframework==3.16.1
frozenlist==1.6.0
idna==3.10
jalali_core==1.0.0
jdatetime==5.2.0
multidict==6.4.3
mysqlclient==2.2.7
pillow==12.0.0
polib==1.2.0
propcache==0.3.1
requests==2.32.3
setuptools==80.4.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.4.0
yarl==1.20.0
